Use bstree function from week 1 test 2
According to https://github.com/mtrebi/memory-allocators,
a free list allocator can reduce the complexity of finding
available memory blocks to O(log n) by introducing
a search tree.
This repository mentions that a red-black tree is
commonly used for maintaining the structure.

I first attempted to integrate the functions provided in
 https://hackmd.io/@sysprog/linux2025-quiz1,
and supplemented them with some necessary features to
 enable the binary search tree to support search,
insertion, and deletion of specific nodes.